requirementsLinux.txt
pillow
numpy
tqdm
python3-fontforge
